more

+++ tests/cn_vip_testsuite/pointer_copy_user_ctrlflow_bitwise.annot.c
@@ -19,11 +19,11 @@ int main()
   /*CN_VIP*/int *q = NULL;
   /*CN_VIP*/bit=0;
   for (int k=0; k<uintptr_t_width; k++)
-  /*@ inv i == (u64) p;
+  /*@ inv i == (integer) p;
           ptr_eq(p, &x);
-          uintptr_t_width == 64u64;
-          (0i32 <= k) && (k <= 64i32);
-          let k_mask = shift_left(1u64, (u64) k) - 1u64;
+          uintptr_t_width == 64;
+          (0 <= k) && (k <= 64);
+          let k_mask = shift_left(1, k) - 1u64;
           j == i & k_mask;
   @*/
   {
@@ -40,5 +40,5 @@ int main()
 #endif
   *q = 11; // CN VIP UB (no annot)
   //CN_VIP printf("*p=%d  *q=%d\n",*p,*q);
-  /*CN_VIP*//*@ assert(*p == 11i32 && *q == 11i32); @*/
+  /*CN_VIP*//*@ assert(*p == 11 && *q == 11); @*/
 }
